prefetch for settings page
